pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "BaiKongQue",email = "2lperkins@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-slim (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pypdf (>=5.5.0,<6.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "chromadb (>=1.0.10,<2.0.0)",
    "openai (>=1.82.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "ollama (>=0.5.1,<0.6.0)"
]

[tool.poetry]
packages = [{include = "src", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
